backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.23
shapely==2.0.6
geojson==3.1.0
python-multipart==0.0.6
pydantic==2.10.3
httpx
numpy<2

# AWS S3 for GeoJSON hosting
boto3==1.34.0

# Wildfire prediction dependencies
requests==2.31.0
schedule==1.2.0

# LLM and AI dependencies
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-core==0.1.9

# RAG and Vector Database
pinecone-client==5.0.0
sentence-transformers==2.2.2

# PDF Generation and Document Processing
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
weasyprint==60.2

# Utility libraries
python-dotenv==1.0.0
jinja2==3.1.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Chatbot dependencies
websockets==12.0

# Satellite imagery comparison
planet==2.4.0
sentinelhub==3.10.0
Pillow==10.0.0
